+++ include/H2MM.h
@@ -123,7 +123,7 @@ public:
      */
     void set_bursts_from_tttr(
         std::shared_ptr<TTTR> tttr,
-        const std::vector<int64_t>& bursts,
+        const std::vector<long long>& bursts,
         const std::vector<std::shared_ptr<Channel>>& stream_channels,
         int min_photons = 3,
         long long time_scale = 1
